refactor(fricon-ui): reorganize chart data imports and types
