+++ crates/fricon-ui/src/features/charts/transform.rs
@@ -6,4 +6,3 @@ pub(crate) mod scatter;
 pub(crate) use self::{
     heatmap::build_heatmap_series, line::build_line_series, scatter::build_scatter_series,
 };
-pub(crate) use crate::features::charts::types::*;
